requirements.txt
selenium~=4.25.0
pytest-html==4.1.1
pytest-xdist==3.6.1
pytest~=8.3.3
requests~=2.32.3
behave==1.2.6
pyyaml
pandas
# cx-Oracle~=8.3.0